Factor out utils, pass balance to Lock
