+++ unlock-app/src/__tests__/components/interface/checkout/Lock.test.tsx
@@ -1,66 +1,13 @@
 import React from 'react'
 import * as rtl from '@testing-library/react'
-import {
-  lockKeysAvailable,
-  lockTickerSymbol,
-  Lock,
-} from '../../../../components/interface/checkout/Lock'
+import { Lock } from '../../../../components/interface/checkout/Lock'
 import * as usePurchaseKey from '../../../../hooks/usePurchaseKey'
 
+const balances = {
+  eth: '500.00',
+}
+
 describe('Checkout Lock', () => {
-  describe('helpers', () => {
-    describe('lockKeysAvailable', () => {
-      it('returns Unlimited if it has unlimited keys', () => {
-        expect.assertions(1)
-        const lock = {
-          unlimitedKeys: true,
-          maxNumberOfKeys: -1,
-          outstandingKeys: 100,
-        }
-        expect(lockKeysAvailable(lock)).toEqual('Unlimited')
-      })
-
-      it('returns the difference between max and outstanding ottherwise', () => {
-        expect.assertions(1)
-        const lock = {
-          maxNumberOfKeys: 203,
-          outstandingKeys: 100,
-        }
-        expect(lockKeysAvailable(lock)).toEqual('103')
-      })
-    })
-
-    describe('lockTickerSymbol', () => {
-      it('returns ETH when it is an ETH lock', () => {
-        expect.assertions(1)
-        const lock = {
-          currencyContractAddress: null,
-        }
-
-        expect(lockTickerSymbol(lock)).toEqual('ETH')
-      })
-
-      it('returns DAI when it is a DAI lock', () => {
-        expect.assertions(1)
-        const lock = {
-          currencyContractAddress: '0xDAI',
-          currencySymbol: 'DAI',
-        }
-
-        expect(lockTickerSymbol(lock)).toEqual('DAI')
-      })
-
-      it('returns ERC20 when it is an unknown ERC20 lock', () => {
-        expect.assertions(1)
-        const lock = {
-          currencyContractAddress: '0xDAI',
-        }
-
-        expect(lockTickerSymbol(lock)).toEqual('ERC20')
-      })
-    })
-  })
-
   describe('Lock', () => {
     let purchaseKey: () => void
     let setPurchasingLockAddress: () => void
@@ -91,6 +38,7 @@ describe('Checkout Lock', () => {
           lock={lock}
           purchasingLockAddress={null}
           setPurchasingLockAddress={setPurchasingLockAddress}
+          balances={balances}
         />
       )
 
@@ -118,6 +66,7 @@ describe('Checkout Lock', () => {
           lock={lock}
           purchasingLockAddress="0xneato"
           setPurchasingLockAddress={setPurchasingLockAddress}
+          balances={balances}
         />
       )
 
@@ -145,6 +94,7 @@ describe('Checkout Lock', () => {
           lock={lock}
           purchasingLockAddress="0xneato"
           setPurchasingLockAddress={setPurchasingLockAddress}
+          balances={balances}
         />
       )
 
@@ -167,6 +117,7 @@ describe('Checkout Lock', () => {
           lock={lock}
           purchasingLockAddress="0xlockaddress"
           setPurchasingLockAddress={setPurchasingLockAddress}
+          balances={balances}
         />
       )
 
@@ -198,6 +149,7 @@ describe('Checkout Lock', () => {
           lock={lock}
           purchasingLockAddress="0xlockaddress"
           setPurchasingLockAddress={setPurchasingLockAddress}
+          balances={balances}
         />
       )
