+++ unlock-app/src/components/interface/checkout/Lock.tsx
@@ -1,6 +1,10 @@
 import React from 'react'
-import { RawLock } from '../../../unlockTypes'
+import { RawLock, Balances } from '../../../unlockTypes'
 import { durationsAsTextFromSeconds } from '../../../utils/durations'
+import {
+  lockKeysAvailable,
+  lockTickerSymbol,
+} from '../../../utils/checkoutLockUtils'
 import { usePurchaseKey } from '../../../hooks/usePurchaseKey'
 import * as LockVariations from './LockVariations'
 
@@ -8,39 +12,7 @@ interface LockProps {
   lock: RawLock
   purchasingLockAddress: string | null
   setPurchasingLockAddress: (lockAddress: string) => void
-}
-
-interface LockKeysAvailableLock {
-  unlimitedKeys?: boolean
-  maxNumberOfKeys?: number
-  outstandingKeys?: number
-}
-
-export const lockKeysAvailable = ({
-  unlimitedKeys,
-  maxNumberOfKeys,
-  outstandingKeys,
-}: LockKeysAvailableLock) => {
-  if (unlimitedKeys) {
-    return 'Unlimited'
-  }
-
-  // maxNumberOfKeys and outstandingKeys are assumed to be defined
-  // if they are ever not, a runtime error can occur
-  return (maxNumberOfKeys! - outstandingKeys!).toLocaleString()
-}
-
-interface LockTickerSymbolLock {
-  currencyContractAddress: string | null
-  currencySymbol?: string
-}
-
-export const lockTickerSymbol = (lock: LockTickerSymbolLock) => {
-  if (lock.currencyContractAddress) {
-    // TODO: if there is no symbol, we probably need something better than "ERC20"
-    return (lock as any).currencySymbol || 'ERC20'
-  }
-  return 'ETH'
+  balances: Balances
 }
 
 export const Lock = ({
